# PERF: fix clean_index_list perf